requirements.txt
google-cloud-dialogflow==2.31.0
python-telegram-bot==21.4
vkbottle==4.3.12
environs==11.0.0
